2008-03-10  Dave Brolley  <brolley@redhat.com>
        PR5189
        * vsprintf.c (_stp_vsnprintf): Arguments for dynamic width and precision
        are of type 'int' again.
        * loc2c-runtime.h (deref_string): Copy the data only if dst is not NULL.
        (deref_buffer): New macro.

+++ runtime/vsprintf.c
@@ -180,7 +180,7 @@ int _stp_vsnprintf(char *buf, size_t size, const char *fmt, va_list args)
 		else if (*fmt == '*') {
 			++fmt;
 			/* it's the next argument */
-			field_width = va_arg(args, int64_t);
+			field_width = va_arg(args, int);
 			if (field_width < 0) {
 				field_width = -field_width;
 				flags |= STP_LEFT;
@@ -196,7 +196,7 @@ int _stp_vsnprintf(char *buf, size_t size, const char *fmt, va_list args)
 			else if (*fmt == '*') {
 				++fmt;
 				/* it's the next argument */
-				precision = va_arg(args, int64_t);
+				precision = va_arg(args, int);
 			}
 			if (precision < 0)
 				precision = 0;
